Bug 50579 - Error count is long, sample count is int. Changed sample count to long.
git-svn-id: https://svn.apache.org/repos/asf/jakarta/jmeter/trunk@1059242 13f79535-47bb-0310-9956-ffa450edef68

+++ src/jorphan/org/apache/jorphan/math/StatCalculator.java
@@ -46,7 +46,7 @@ public abstract class StatCalculator<T extends Number & Comparable<? super T>> {
 
     private double deviation = 0;
 
-    private int count = 0;
+    private long count = 0;
 
     private T min;
 
@@ -187,7 +187,7 @@ public abstract class StatCalculator<T extends Number & Comparable<? super T>> {
         return max;
     }
 
-    public int getCount() {
+    public long getCount() {
         return count;
     }
